--- src/pages/AuthCallback/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AuthCallback';
